More work done on leave-application.

+++ ums-api/src/main/java/org/ums/domain/model/immutable/common/LmsAppStatus.java
@@ -14,7 +14,7 @@ import java.util.Date;
  * Created by Monjur-E-Morshed on 06-May-17.
  */
 public interface LmsAppStatus extends Serializable, LastModifier, EditType<MutableLmsAppStatus>, Identifier<Integer> {
-  int getLmsAppId();
+  Long getLmsAppId();
 
   LmsApplication getLmsApplication();
